requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy[asyncio]
aiosqlite
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
python-multipart